Factor out common stream utils
