+++ jumpstarter/driver/base.py
@@ -26,10 +26,8 @@ from jumpstarter.common.streams import (
     DriverStreamRequest,
     ResourceStreamRequest,
     StreamRequest,
-    create_memory_stream,
-    forward_stream,
 )
-from jumpstarter.streams import RouterStream
+from jumpstarter.streams import RouterStream, create_memory_stream, forward_stream
 from jumpstarter.v1 import jumpstarter_pb2, jumpstarter_pb2_grpc, router_pb2_grpc
 
 from .decorators import (
