+++ jumpstarter/streams/__init__.py
@@ -1,3 +1,4 @@
+from .common import create_memory_stream, forward_stream
 from .router import RouterStream
 
-__all__ = ["RouterStream"]
+__all__ = ["create_memory_stream", "forward_stream", "RouterStream"]
